Add files via upload
Below is the website for the "Reeborg's World" robot game, where you direct a robot to the finish line using Python commands. The attached file includes solutions to the following challenges on the site: "Hurdle 1–4", and "Maze". It uses simple if/elif/else logic and while loops with pre-built commands for the game. The attached file is not a py file, just the code used in the python console on the website. 

https://reeborg.ca/reeborg.html?lang=en&mode=python&menu=worlds%2Fmenus%2Freeborg_intro_en.json&name=Hurdle%201&url=worlds%2Ftutorial_en%2Fhurdle1.json